chore: split fabric gen helper function
